Move `URLSessionHTTPClient` to production
