+++ CEPSearch/CEPSearchTests/CEP API/URLSessionHTTPClientTests.swift
@@ -16,30 +16,6 @@ import CEPSearch
  https://github.com/igdutra/NASA-Gallery/blob/main/NASAGallery/NASAGalleryTests/Gallery%20API/URLSessionHTTPClientTests.swift
 */
 
-public final class URLSessionHTTPClient: HTTPClient {
-    
-    private let session: URLSession
-    
-    public init(session: URLSession = .shared) {
-        self.session = session
-    }
-    
-    public func getData(from url: URL) async throws -> (data: Data, response: HTTPURLResponse) {
-        do {
-            let (data, response) = try await session.data(from: url)
-            
-            guard let httpResponse = response as? HTTPURLResponse else {
-                throw URLError(.cannotParseResponse)
-            }
-            
-            return (data: data, response: httpResponse)
-        } catch {
-            throw error
-        }
-    }
-}
-
-
 final class URLSessionHTTPClientTests: XCTestCase {
     
     // MARK: - SetUp & TearDown
